Fixed BraveTabGroupHeaderTest.TitleChipClipPathClearedInVerticalTabs
