+++ browser/ui/views/tabs/BUILD.gn
@@ -55,6 +55,8 @@ source_set("unit_tests") {
     "brave_tab_strip_layout_helper_unittest.cc",
     "brave_tab_strip_unittest.cc",
     "brave_tab_unittest.cc",
+    "mock_browser_window_interface_with_vertical_tab_controller.cc",
+    "mock_browser_window_interface_with_vertical_tab_controller.h",
     "tab_style_views_unittest.cc",
   ]
