+++ browser/ui/views/tabs/brave_tab_unittest.cc
@@ -8,16 +8,13 @@
 #include <algorithm>
 
 #include "base/test/scoped_feature_list.h"
-#include "brave/browser/ui/focus_mode/focus_mode_controller.h"
 #include "brave/browser/ui/tabs/brave_tab_prefs.h"
-#include "brave/browser/ui/tabs/public/vertical_tab_controller.h"
 #include "brave/browser/ui/views/tabs/brave_tab_strip_layout_helper.h"
+#include "brave/browser/ui/views/tabs/mock_browser_window_interface_with_vertical_tab_controller.h"
 #include "brave/components/tabs/public/tree_tab_node.h"
 #include "brave/components/tabs/public/tree_tab_node_id.h"
 #include "brave/components/tabs/public/tree_tab_node_tab_collection.h"
-#include "chrome/browser/ui/browser_window/public/browser_window_features.h"
 #include "chrome/browser/ui/browser_window/public/browser_window_interface.h"
-#include "chrome/browser/ui/browser_window/test/mock_browser_window_interface.h"
 #include "chrome/browser/ui/layout_constants.h"
 #include "chrome/browser/ui/tabs/features.h"
 #include "chrome/browser/ui/tabs/tab_style.h"
@@ -61,24 +58,6 @@ class MockTabSlotController : public FakeTabSlotController {
               GetTreeTabNode,
               (const tree_tab::TreeTabNodeId& id),
               (const, override));
-};
-
-class MockBrowserWindowInterfaceWithVerticalTabController
-    : public MockBrowserWindowInterface {
- public:
-  explicit MockBrowserWindowInterfaceWithVerticalTabController(
-      PrefService* prefs) {
-    features_.SetVerticalTabControllerForTesting(
-        std::make_unique<VerticalTabController>(
-            BrowserWindowInterface::TYPE_NORMAL, prefs,
-            &focus_mode_controller_));
-    ON_CALL(*this, GetFeatures()).WillByDefault(testing::ReturnRef(features_));
-  }
-  ~MockBrowserWindowInterfaceWithVerticalTabController() override = default;
-
- private:
-  FocusModeController focus_mode_controller_;
-  BrowserWindowFeatures features_;
 };
 
 class BraveTabTest : public ChromeViewsTestBase {
